@@ -17,7 +17,7 @@ package deployment
 import (
 	"context"
 	"crypto/tls"
-	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"net/http"
@@ -33,7 +33,6 @@ import (
 	"github.com/docker/docker/api/types/volume"
 	"github.com/docker/docker/client"
 	"github.com/hashicorp/go-version"
-	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
 	"github.com/valyala/fasthttp"
 
@@ -175,8 +174,10 @@ func (p *PMM) ClickhouseURL() string {
 		p.t.Fatal(err)
 	}
 
-	u.User = pkgUtil.GetClickhouseUser(p.GetFullVersionString())
-
+	u.User = url.UserPassword("default", "clickhouse")
+	if pkgUtil.CheckVer(p.GetVersion(), "<= 3.1.0") {
+		u.User = nil
+	}
 	u.Scheme = "clickhouse"
 	u.Path = "pmm"
 	if strings.Contains(u.Host, ":") {
@@ -208,7 +209,7 @@ func (pmm *PMM) Deploy(ctx context.Context) error {
 		return nil
 	}
 	if err := pmm.deploy(ctx); err != nil {
-		return errors.Wrap(err, "failed to deploy")
+		return fmt.Errorf("failed to deploy: %w", err)
 	}
 	*pmm.deployed = true
 	pmm.deployedMu.Unlock()
@@ -234,43 +235,6 @@ func (pmm *PMM) GetVersion() *version.Version {
 	}
 }
 
-// Return the full version asking it to the PMM server itself.
-func (pmm *PMM) GetFullVersionString() string {
-	type ServerVersion struct {
-		Version string `json:"version"`
-	}
-
-	type versionResp struct {
-		Version string        `json:"version"`
-		Server  ServerVersion `json:"server"`
-	}
-
-	client, err := pmm.NewClient()
-
-	pmmVersionURL := pmm.PMMURL() + "/v1/version"
-
-	if err != nil {
-		pmm.t.Fatal(err)
-	}
-
-	response_code, response, err := client.Get(pmmVersionURL)
-
-	if err != nil {
-		pmm.t.Fatal(err)
-	}
-
-	if response_code != fasthttp.StatusOK {
-		pmm.t.Fatal(fmt.Errorf("non-ok status: %d", response_code))
-	}
-
-	var versionInfo versionResp
-	if err := json.Unmarshal(response, &versionInfo); err != nil {
-		pmm.t.Fatal(err)
-	}
-
-	return versionInfo.Server.Version
-}
-
 func (pmm *PMM) SetVersion(version string) {
 	pmm.pmmVersion = version
 }
@@ -294,13 +258,13 @@ var checkImagesMu sync.Mutex
 func (pmm *PMM) deploy(ctx context.Context) error {
 	dockerCli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
-		return errors.Wrap(err, "failed to create docker client")
+		return fmt.Errorf("failed to create docker client: %w", err)
 	}
 	defer dockerCli.Close() //nolint:errcheck
 
 	pmm.Log("Destroying existing deployment")
 	if err := destroy(ctx, filters.NewArgs(filters.Arg("label", PerconaLabel+"="+pmm.testName)), pmm); err != nil {
-		return errors.Wrap(err, "failed to destroy existing deployment")
+		return fmt.Errorf("failed to destroy existing deployment: %w", err)
 	}
 
 	pmm.Log("Checking images")
@@ -309,7 +273,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 		exists, err := ImageExists(ctx, image)
 		if err != nil {
 			checkImagesMu.Unlock()
-			return errors.Wrap(err, "failed to check image")
+			return fmt.Errorf("failed to check image: %w", err)
 		}
 		if exists {
 			pmm.Log("Image", image, "exists")
@@ -318,7 +282,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 		pmm.Log("Pulling image", image)
 		if err := PullImage(ctx, image); err != nil {
 			checkImagesMu.Unlock()
-			return errors.Wrap(err, "failed to pull image")
+			return fmt.Errorf("failed to pull image: %w", err)
 		}
 	}
 	checkImagesMu.Unlock()
@@ -341,7 +305,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 	if err := util.RetryOnError(tCtx, func() error {
 		_, err := dockerCli.NetworkInspect(ctx, netresp.ID, network.InspectOptions{})
 		if err != nil {
-			return errors.Wrapf(err, "failed to inspect network %s", netresp.ID)
+			return fmt.Errorf("failed to inspect network %s: %w", netresp.ID, err)
 		}
 		return nil
 	}); err != nil {
@@ -350,17 +314,17 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 
 	pmm.Log("Creating PMM server")
 	if err := pmm.CreatePMMServer(ctx, dockerCli, netresp.ID); err != nil {
-		return errors.Wrap(err, "failed to create pmm server")
+		return fmt.Errorf("failed to create pmm server: %w", err)
 	}
 
 	pmm.Log("Creating PMM client")
 	if err := pmm.CreatePMMClient(ctx, dockerCli, netresp.ID); err != nil {
-		return errors.Wrap(err, "failed to create pmm client")
+		return fmt.Errorf("failed to create pmm client: %w", err)
 	}
 
 	pmm.Log("Creating mongo")
 	if err := pmm.CreateMongo(ctx, dockerCli, netresp.ID); err != nil {
-		return errors.Wrap(err, "failed to create mongo container")
+		return fmt.Errorf("failed to create mongo container: %w", err)
 	}
 
 	pmm.Log("Waiting for mongo to be ready")
@@ -371,7 +335,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 		return pmm.PingMongo(ctx)
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to connect to mongo")
+		return fmt.Errorf("failed to connect to mongo: %w", err)
 	}
 
 	pmm.Log("Adding mongo to PMM")
@@ -385,7 +349,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 			"mongo",
 			pmm.MongoContainerName()+":27017")
 	}); err != nil {
-		return errors.Wrap(err, "failed to add mongo to PMM")
+		return fmt.Errorf("failed to add mongo to PMM: %w", err)
 	}
 
 	pmm.Log("Ping clickhouse with driver")
@@ -394,7 +358,7 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 	if err := util.RetryOnError(tCtx, func() error {
 		return pmm.PingClickhouse(ctx)
 	}); err != nil {
-		return errors.Wrap(err, "failed to ping clickhouse")
+		return fmt.Errorf("failed to ping clickhouse: %w", err)
 	}
 
 	return nil
@@ -403,28 +367,28 @@ func (pmm *PMM) deploy(ctx context.Context) error {
 func (pmm *PMM) Restart(ctx context.Context) error {
 	dockerCli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
-		return errors.Wrap(err, "failed to create docker client")
+		return fmt.Errorf("failed to create docker client: %w", err)
 	}
 	defer dockerCli.Close() //nolint:errcheck
 
 	if err := dockerCli.ContainerRestart(ctx, *pmm.pmmServerContainerID, container.StopOptions{
 		Timeout: nil, // 10 seconds
 	}); err != nil {
-		return errors.Wrap(err, "failed to restart pmm server")
+		return fmt.Errorf("failed to restart pmm server: %w", err)
 	}
 	if err := pmm.SetServerPublishedPorts(ctx, dockerCli); err != nil {
-		return errors.Wrap(err, "failed to set server published ports")
+		return fmt.Errorf("failed to set server published ports: %w", err)
 	}
 
 	tCtx, cancel := context.WithTimeout(ctx, getTimeout)
 	defer cancel()
-	if pkgUtil.CheckIsVer2(pmm.GetVersion()) {
+	if pkgUtil.CheckVer(pmm.GetVersion(), "< 3.0.0") {
 		if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/version"); err != nil && !errors.Is(err, io.EOF) {
-			return errors.Wrap(err, "failed to ping PMM")
+			return fmt.Errorf("failed to ping PMM: %w", err)
 		}
 	} else {
 		if err := getUntilOk(tCtx, pmm.PMMURL()+"/v1/server/version"); err != nil && !errors.Is(err, io.EOF) {
-			return errors.Wrap(err, "failed to ping PMM")
+			return fmt.Errorf("failed to ping PMM: %w", err)
 		}
 	}
 	return nil
@@ -469,7 +433,7 @@ func (l *defaultLogger) Log(args ...any) {
 func destroy(ctx context.Context, filters filters.Args, log logger) error {
 	dockerCli, err := client.NewClientWithOpts(client.FromEnv, client.WithAPIVersionNegotiation())
 	if err != nil {
-		return errors.Wrap(err, "failed to create docker client")
+		return fmt.Errorf("failed to create docker client: %w", err)
 	}
 	defer dockerCli.Close() //nolint:errcheck
 
@@ -478,7 +442,7 @@ func destroy(ctx context.Context, filters filters.Args, log logger) error {
 		Filters: filters,
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to list containers")
+		return fmt.Errorf("failed to list containers: %w", err)
 	}
 
 	zero := 0
@@ -498,7 +462,7 @@ func destroy(ctx context.Context, filters filters.Args, log logger) error {
 		Filters: filters,
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to list volumes")
+		return fmt.Errorf("failed to list volumes: %w", err)
 	}
 	for _, vol := range volumes.Volumes {
 		if err := dockerCli.VolumeRemove(ctx, vol.Name, true); err != nil {
@@ -510,7 +474,7 @@ func destroy(ctx context.Context, filters filters.Args, log logger) error {
 		Filters: filters,
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to list networks")
+		return fmt.Errorf("failed to list networks: %w", err)
 	}
 
 	for _, n := range networks {
@@ -540,7 +504,7 @@ func (p *PMM) NewClient() (*grafanaClient.Client, error) {
 	}
 	grafanaClient, err := grafanaClient.NewClient(httpC, authParams)
 	if err != nil {
-		return nil, errors.Wrap(err, "new client")
+		return nil, fmt.Errorf("new client: %w", err)
 	}
 	return grafanaClient, nil
 }
