@@ -16,21 +16,21 @@ package transferer
 
 import (
 	"archive/tar"
-	"compress/gzip"
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"os"
 	"path"
 	"time"
 
-	"github.com/pkg/errors"
 	"github.com/rs/zerolog/log"
 
 	"pmm-dump/pkg/dump"
+	"pmm-dump/pkg/encryption"
 )
 
-func ReadMetaFromDump(dumpPath string, piped bool) (*dump.Meta, error) {
+func ReadMetaFromDump(dumpPath string, piped bool, e encryption.Options) (*dump.Meta, error) {
 	var file *os.File
 	if piped {
 		file = os.Stdin
@@ -38,18 +38,17 @@ func ReadMetaFromDump(dumpPath string, piped bool) (*dump.Meta, error) {
 		var err error
 		file, err = os.Open(dumpPath) //nolint:gosec
 		if err != nil {
-			return nil, errors.Wrap(err, "failed to open file")
+			return nil, fmt.Errorf("failed to open file: %w", err)
 		}
 	}
 	defer file.Close() //nolint:errcheck
 
-	gzr, err := gzip.NewReader(file)
+	r, err := dump.NewReader(file, &e)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to open as gzip")
+		return nil, fmt.Errorf("failed to create readers: %w", err)
 	}
-	defer gzr.Close() //nolint:errcheck
-
-	tr := tar.NewReader(gzr)
+	defer r.Close() //nolint:errcheck
+	tr := r.GetTarReader()
 
 	for {
 		log.Debug().Msg("Reading files from dump...")
@@ -62,7 +61,7 @@ func ReadMetaFromDump(dumpPath string, piped bool) (*dump.Meta, error) {
 		}
 
 		if err != nil {
-			return nil, errors.Wrap(err, "failed to read a file from dump")
+			return nil, fmt.Errorf("failed to read a file from dump: %w", err)
 		}
 
 		_, filename := path.Split(header.Name)
@@ -75,7 +74,7 @@ func ReadMetaFromDump(dumpPath string, piped bool) (*dump.Meta, error) {
 
 		meta, err := readMetafile(tr)
 		if err != nil {
-			return nil, errors.Wrap(err, "failed to read meta file")
+			return nil, fmt.Errorf("failed to read meta file: %w", err)
 		}
 
 		return meta, nil
@@ -98,11 +97,11 @@ func writeMetafile(tw *tar.Writer, meta dump.Meta) error {
 		ModTime:  time.Now(),
 	})
 	if err != nil {
-		return errors.Wrap(err, "failed to write dump meta")
+		return fmt.Errorf("failed to write dump meta: %w", err)
 	}
 
 	if _, err = tw.Write(metaContent); err != nil {
-		return errors.Wrap(err, "failed to write dump meta content")
+		return fmt.Errorf("failed to write dump meta content: %w", err)
 	}
 
 	return nil
@@ -111,12 +110,12 @@ func writeMetafile(tw *tar.Writer, meta dump.Meta) error {
 func readMetafile(r io.Reader) (*dump.Meta, error) {
 	metaBytes, err := io.ReadAll(r)
 	if err != nil {
-		return nil, errors.Wrap(err, "failed to read bytes")
+		return nil, fmt.Errorf("failed to read bytes: %w", err)
 	}
 
 	var meta dump.Meta
 	if err := json.Unmarshal(metaBytes, &meta); err != nil {
-		return nil, errors.Wrap(err, "failed to unmarshal")
+		return nil, fmt.Errorf("failed to unmarshal: %w", err)
 	}
 
 	return &meta, nil
