@@ -50,7 +50,7 @@ func TestExportImport(t *testing.T) {
 
 	pmm.Log("Checking filtering with `--instance` flag")
 	args := []string{"-d", filepath.Join(testDir, "filter-dump.tar.gz"), "--pmm-url", pmm.PMMURL(), "--dump-qan", "--click-house-url", pmm.ClickhouseURL(), "--instance", "pmm-server", "-v"}
-	stdout, stderr, err := b.Run(append([]string{"export", "--ignore-load"}, args...)...)
+	stdout, stderr, err := b.Run(append([]string{"export", "--ignore-load", "--no-encryption"}, args...)...)
 	if err != nil {
 		t.Fatal("failed to export", err, stdout, stderr)
 	}
@@ -58,12 +58,12 @@ func TestExportImport(t *testing.T) {
 
 	pmm.Log("Checking export with `--export-services-info` flag")
 	args = []string{"-d", filepath.Join(testDir, "export-services-info.tar.gz"), "--pmm-url", pmm.PMMURL(), "--dump-core", "--click-house-url", pmm.ClickhouseURL(), "--export-services-info", "-v"}
-	stdout, stderr, err = b.Run(append([]string{"export", "--ignore-load"}, args...)...)
+	stdout, stderr, err = b.Run(append([]string{"export", "--ignore-load", "--no-encryption"}, args...)...)
 	if err != nil {
 		t.Fatal("failed to export", err, stdout, stderr)
 	}
 
-	stdout, stderr, err = b.Run("show-meta", "-d", filepath.Join(testDir, "export-services-info.tar.gz"))
+	stdout, stderr, err = b.Run("show-meta", "--no-encryption", "-d", filepath.Join(testDir, "export-services-info.tar.gz"))
 	if err != nil {
 		t.Fatal(err, stdout, stderr)
 	}
@@ -81,13 +81,13 @@ func TestExportImport(t *testing.T) {
 	args = []string{"-d", filepath.Join(testDir, "dump.tar.gz"), "--pmm-url", pmm.PMMURL(), "--dump-qan", "--click-house-url", pmm.ClickhouseURL(), "-v"}
 
 	pmm.Log("Exporting data to", filepath.Join(testDir, "dump.tar.gz"))
-	stdout, stderr, err = b.Run(append([]string{"export", "--ignore-load"}, args...)...)
+	stdout, stderr, err = b.Run(append([]string{"export", "--ignore-load", "--no-encryption"}, args...)...)
 	if err != nil {
 		t.Fatal("failed to export", err, stdout, stderr)
 	}
 	args = []string{"-d", filepath.Join(testDir, "dump.tar.gz"), "--pmm-url", newPMM.PMMURL(), "--dump-qan", "--click-house-url", newPMM.ClickhouseURL(), "-v"}
 	pmm.Log("Importing data from", filepath.Join(testDir, "dump.tar.gz"))
-	stdout, stderr, err = b.Run(append([]string{"import"}, args...)...)
+	stdout, stderr, err = b.Run(append([]string{"import", "--no-encryption"}, args...)...)
 	if err != nil {
 		t.Fatal("failed to import", err, stdout, stderr)
 	}
@@ -95,7 +95,7 @@ func TestExportImport(t *testing.T) {
 
 func TestShowMeta(t *testing.T) {
 	var b util.Binary
-	stdout, stderr, err := b.Run("show-meta", "-d", filepath.Join(util.RepoPath, "internal", "test", "e2e", "testdata", "onlymeta.tar.gz"))
+	stdout, stderr, err := b.Run("show-meta", "--no-encryption", "-d", filepath.Join(util.RepoPath, "internal", "test", "e2e", "testdata", "onlymeta.tar.gz"))
 	if err != nil {
 		t.Fatal(err, stdout, stderr)
 	}
