@@ -20,6 +20,7 @@ import (
 	"time"
 
 	"pmm-dump/pkg/dump"
+	"pmm-dump/pkg/encryption"
 )
 
 func TestExport(t *testing.T) {
@@ -127,6 +128,12 @@ func TestExport(t *testing.T) {
 					workersCount: opt.workersCount,
 					file:         bytes.NewBuffer(nil),
 				}
+				e := encryption.Options{
+					Encryption: false,
+					JustKey:    false,
+					Pass:       "",
+					Filepath:   "",
+				}
 				var meta dump.Meta
 				var chunks []dump.ChunkMeta
 				if tt.chunkSourceType != dump.UndefinedSource {
@@ -140,7 +147,7 @@ func TestExport(t *testing.T) {
 				if err != nil {
 					t.Fatal(err, "failed to create new chunk pool")
 				}
-				err = tr.Export(ctx, fakeStatusGetter{status: tt.loadStatus.status, waitCount: tt.loadStatus.waitCount, statusAfterWait: tt.loadStatus.statusAfterWait, count: new(int)}, meta, pool, new(bytes.Buffer))
+				err = tr.Export(ctx, fakeStatusGetter{status: tt.loadStatus.status, waitCount: tt.loadStatus.waitCount, statusAfterWait: tt.loadStatus.statusAfterWait, count: new(int)}, meta, pool, new(bytes.Buffer), e)
 				if err != nil {
 					if tt.shouldErr {
 						return
