@@ -26,6 +26,7 @@ import (
 	"time"
 
 	"pmm-dump/pkg/dump"
+	"pmm-dump/pkg/encryption"
 )
 
 func TestImport(t *testing.T) {
@@ -127,8 +128,14 @@ func TestImport(t *testing.T) {
 					workersCount: opt.workersCount,
 					file:         buf,
 				}
+				e := encryption.Options{
+					Encryption: false,
+					JustKey:    false,
+					Pass:       "",
+					Filepath:   "",
+				}
 				meta := dump.Meta{}
-				err := tr.Import(ctx, meta)
+				err := tr.Import(ctx, meta, e)
 				if err != nil {
 					if tt.shouldErr {
 						return
